Initial vote counting logic No soft fork is yet implemented, this just counts votes and balances. New tables:   * votebalance -  key: [address][0 u32]         value: repeated: [block_number u32][balance u64]                    key: [address][block_num u32] value: [is_candidate u8][vote for address]   * votewinnerdb - key: [block_num u32]          value: [vb_hash [32]u8][winner address] The votebalance table contains both votes and balance info. The prefix of every key entry is an address in pkScript format as it would appear in a TX output. Following that is 4 bytes, if those 4 bytes are all zeros then it is a balance entry, if the 4 bytes are non-zero then it is a vote.
Balance:
The balance entry value is a repeated array of block number + balance entries. Block numbers are little
endian uint32 and balance entries are little endian uint64s. Multiple balance entries are stored in
order to keep the balance at time of a vote epoch, and also to have the balance at the current moment
which is needed in order to correctly compute the balance at the next vote epoch. The balance table is
able to correctly handle a rollback of up to 1 epoch worth of blocks (1 week) without failing. More
that and it will not correctly know the balance at the time of the previous epoch and thus will fail.
Balances are deleted from the db when they have been zero for more than 2 epochs.

Vote:
The vote entry is prefixed also with the address in pkScript form, this makes all information for one
address fall in the same location in the db. After the address is a BIG ENDIAN number of the block when
the vote took place. The value of a vote begins with 1 byte which indicates whether the voter intends
to candidate and therefore can potentially win. Following that byte is the address which they are
voting for in pkScript form. Votes expire after they are more than 1 year (52 epochs) old, and are thus
deleted from the db when they are more than 52+2 epochs old.

Winner:
The votewinnerdb key is the block number in BIG ENDIAN form and the winner value begins with 32 byte
hash of the votebalance table followed by the address of the winner. The hash of the votebalance table
serves as a checksum to verify the integrity of the table and can be compared between instances to
ensure the correct behavior of the code.

Computation:
This code runs a thread (goroutine) which is notified of each block that is added or removed.
There are a couple of constants which make this logic work:

EpochBlocks = (60 * 24 * 7) -- this is the number of blocks per vote computation.

startCountOffset = 60 -- vote counting does not begin until this many blocks after the epoch boundary
this makes it less likely that the count will need to be restarted because there was a rollback around
the time of the epoch.

InaugurationOffset = 360 -- the vote does not become EFFECTIVE until after this many blocks has passed
since the epoch. This means the vote counting thread can take 300 blocks (about 5 hours) to complete
without interrupting the normal operation of the blockchain.

limitCandidates = 100000 -- the maximum number of addresses that can be voted for. If there are more
than this number of addresses being voted for, the addresses with lowest balance will be disconsidered.
This limit is imposed to put an upper bound on memory usage.

After block number N passes where N % EpochBlocks = startCountOffset, the vote counting thread begins
operation. The counting thread scans the votebalances table taking into account most recent vote and
balance as of block N - startCountOffset (i.e. the epoch boundary). This scanning is done in steps so
as not to hold the database lock for too long. It is clear that no change can happen to balances or
votes that are startCountOffset old, so there is no reason to hold the database read lock for the
entirety of the computation. If at any time during the computation, a rollback causes N % EpochBlocks
to become less than startCountOffset, the computation is simply aborted and then we wait until
N % EpochBlocks = startCountOffset becomes true again.

Computation takes place first by scanning the votebalance table and collecting all voting addresses
into a redblack tree ordered by address balance. If the tree size goes over limitCandidates then the
poorest addresses are dropped. Then if any addresses had to be dropped, we re-scan the table again
to apply votes from these dropped addresses to the candidates they voted for. In case a voter votes
for a candidate who has also been dropped from the tree, that vote is disregarded. After this, the
votes are passed to Electorium https://github.com/cjdelisle/Electorium_go and then the winner and
the hash of the table is stored with the block number N - startCountOffset in votewinnerdb.

It is important to note that the block number where the vote becomes effective, i.e. the
"inaguration" is equal to N - startCountOffset + InaugurationOffset. So InaugurationOffset must be
added to the number stored in votewinnerdb.